+++ packages/multi-test/src/bank.rs
@@ -346,75 +346,4 @@ mod test {
         let err = bank.execute(&mut store, rcpt, msg).unwrap_err();
         assert!(err.contains("Overflow"));
     }
-
-    fn query_cache(cache: &BankCache, rcpt: &Addr) -> Vec<Coin> {
-        let query = BankQuery::AllBalances {
-            address: rcpt.into(),
-        };
-        let res = cache.query(query).unwrap();
-        let val: AllBalanceResponse = from_slice(&res).unwrap();
-        val.amount
-    }
-
-    fn query_router(cache: &BankRouter, rcpt: &Addr) -> Vec<Coin> {
-        let query = BankQuery::AllBalances {
-            address: rcpt.into(),
-        };
-        let res = cache.query(query).unwrap();
-        let val: AllBalanceResponse = from_slice(&res).unwrap();
-        val.amount
-    }
-
-    #[test]
-    fn multi_level_bank_cache() {
-        let mut store = MockStorage::new();
-
-        let owner = Addr::unchecked("owner");
-        let rcpt = Addr::unchecked("recipient");
-        let init_funds = vec![coin(20, "btc"), coin(100, "eth")];
-
-        // set money
-        let bank = SimpleBank {};
-        bank.set_balance(&mut store, &owner, init_funds).unwrap();
-        let mut router = BankRouter::new(bank, Box::new(store));
-
-        // cache 1 - send some tokens
-        let mut cache = router.cache();
-        let msg = BankMsg::Send {
-            to_address: rcpt.clone().into(),
-            amount: coins(25, "eth"),
-        };
-        cache.execute(owner.clone(), msg).unwrap();
-
-        // shows up in cache
-        let cached_rcpt = query_cache(&cache, &rcpt);
-        assert_eq!(coins(25, "eth"), cached_rcpt);
-        let router_rcpt = query_router(&router, &rcpt);
-        assert_eq!(router_rcpt, vec![]);
-
-        // now, second level cache
-        let mut cache2 = cache.cache();
-        let msg = BankMsg::Send {
-            to_address: rcpt.clone().into(),
-            amount: coins(12, "eth"),
-        };
-        cache2.execute(owner, msg).unwrap();
-
-        // shows up in 2nd cache
-        let cached_rcpt = query_cache(&cache, &rcpt);
-        assert_eq!(coins(25, "eth"), cached_rcpt);
-        let cached2_rcpt = query_cache(&cache2, &rcpt);
-        assert_eq!(coins(37, "eth"), cached2_rcpt);
-
-        // apply second to first
-        let ops = cache2.prepare();
-        ops.commit(&mut cache);
-
-        // apply first to router
-        let ops = cache.prepare();
-        ops.commit(&mut router);
-
-        let committed = query_router(&router, &rcpt);
-        assert_eq!(coins(37, "eth"), committed);
-    }
 }
